requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==5.3.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django>=2.0.10

# Environment
django-environ==0.4.5

# Django REST Framework
<<<<<<< HEAD
djangorestframework==3.12.2

# JWT
pyjwt>=2.0.0
=======
djangorestframework==3.9.0

# JWT
pyjwt==1.7.1
>>>>>>> bbc966a9d58bf236d512eb56eb1ecf2ab5fc9f29

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==4.1.2

# Celery
redis>=3.2.0
django-redis==4.10.0
celery==4.2.1
tornado>=5.0.0,<6.0.0
flower==0.9.2
